removed warning in native image ops

+++ viewer/nativeimageops.hpp
@@ -51,9 +51,9 @@ private:
 		TYPE currentMin = std::numeric_limits<TYPE>::max();
 		#pragma omp parallel for
 
-		for( unsigned short z = start[2] + 1; z < end[2]; z++ ) {
-			for( unsigned short y = start[1] + 1; y < end[1]; y++ ) {
-				for( unsigned short x = start[0] + 1; x < end[0]; x++ ) {
+		for( int32_t z = start[2] + 1; z < end[2]; z++ ) {
+			for( int32_t y = start[1] + 1; y < end[1]; y++ ) {
+				for( int32_t x = start[0] + 1; x < end[0]; x++ ) {
 					currentValue = static_cast<data::Image &>( typedImage ).voxel<TYPE>( x, y, z );
 
 					if( currentValue < currentMin ) {
@@ -89,9 +89,9 @@ private:
 		TYPE currentMax = std::numeric_limits<TYPE>::min();
 		#pragma omp parallel for
 
-		for( unsigned short z = start[2] + 1; z < end[2]; z++ ) {
-			for( unsigned short y = start[1] + 1; y < end[1]; y++ ) {
-				for( unsigned short x = start[0] + 1; x < end[0]; x++ ) {
+		for( int32_t z = start[2] + 1; z < end[2]; z++ ) {
+			for( int32_t y = start[1] + 1; y < end[1]; y++ ) {
+				for( int32_t x = start[0] + 1; x < end[0]; x++ ) {
 					currentValue = static_cast<data::Image &>( typedImage ).voxel<TYPE>( x, y, z );
 
 					if( currentValue > currentMax ) {
